pyproject.toml
[tool.poetry]
name = "minecraft-server-manager"
version = "0.1.0"
description = "Telegram bot for managing Minecraft servers"
readme = "README.md"
license = "MIT"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Telegram Bot
aiogram = "^3.13"

# Async
aiosqlite = "^0.20"
aiofiles = "^24.1"
httpx = "^0.27"

# Data validation & settings
pydantic = "^2.9"
pydantic-settings = "^2.5"
pyyaml = "^6.0"

# Utils
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7"
pytest = "^8.3"
pytest-asyncio = "^0.24"

[tool.poetry.scripts]
mcmanager = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

